pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "artificial-morphogenesis-evo"
version = "1.0.0"
description = "Evolutionary cellular automata for artificial morphogenesis"
readme = "README.md"
requires-python = ">=3.9"

authors = [
  { name = "William Greenwood" },
  { name = "Phil Marino" },
  { name = "Ethan Weathers" }
]

dependencies = [
  "numpy>=1.20",
  "pandas>=1.3",
  "matplotlib>=3.5",
  "scipy>=1.7",
  "PyYAML>=6.0"
]

[tool.setuptools.packages.find]
include = ["src*"]
